# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Motor Ping API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/clients": {
      "get": {
        "tags": [
          "clients"
        ],
        "summary": "Listar",
        "operationId": "listar_clients_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "clients"
        ],
        "summary": "Crear",
        "description": "Dar de alta al cliente. Si su ficha estaba archivada, la revive en vez de duplicarla.\n\nRevivir no es un atajo: el telefono es unico dentro del taller, asi que sin esto\narchivar dejaria ese numero inservible para siempre. Y es lo que el taller espera del\ncliente que vuelve -llega con el historial de su auto, no de cero-.",
        "operationId": "crear_clients_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteEntrada"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/vehicles": {
      "get": {
        "tags": [
          "vehicles"
        ],
        "summary": "Listar",
        "description": "Buscar por `plate` es como se llega a un auto que ya estuvo aca.\n\nLa patente es el unico dato que el mecanico tiene en la mano cuando el auto entra\npor la puerta, y un auto que vuelve es el caso normal de un taller, no la excepcion.\n\nLos autos de un cliente archivado no salen. Si salieran, buscar la patente llevaria a\nuna ficha que el taller ya no ve y armar la orden fallaria sin explicar por que. Por\nid si se pueden seguir pidiendo: es como los muestra una orden vieja.",
        "operationId": "listar_vehicles_get",
        "parameters": [
          {
            "name": "clientId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Clientid"
            }
          },
          {
            "name": "plate",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plate"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "vehicles"
        ],
        "summary": "Crear",
        "operationId": "crear_vehicles_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehiculoEntrada"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/orders": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "Listar",
        "operationId": "listar_orders_get",
        "parameters": [
          {
            "name": "open",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Open"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "Crear",
        "operationId": "crear_orders_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrdenEntrada"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/workshops": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Crear Taller",
        "operationId": "crear_taller_admin_workshops_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AltaTallerEntrada"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Listar Talleres",
        "description": "La lista sirve para una sola pregunta: quien esta usando esto de verdad.\n\nPor eso trae el conteo de ordenes y no solo los datos de contacto. Se arma con dos\nsubconsultas correlacionadas y no con joins: un taller con dos duenos duplicaria filas.",
        "operationId": "listar_talleres_admin_workshops_get",
        "parameters": [
          {
            "name": "archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Archived"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/workshops/{taller_id}": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Editar Taller",
        "operationId": "editar_taller_admin_workshops__taller_id__patch",
        "parameters": [
          {
            "name": "taller_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Taller Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TallerEdicion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Dar De Baja",
        "description": "El taller se fue. Sale de la lista y nadie de ahi entra, pero no se borra nada.\n\nSus ordenes, sus clientes y su historial por patente quedan enteros: si vuelve en dos\nmeses, `restore` le devuelve todo donde lo dejo. Es la misma decision que con los\nclientes y las ordenes, y por la misma razon.",
        "operationId": "dar_de_baja_admin_workshops__taller_id__delete",
        "parameters": [
          {
            "name": "taller_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Taller Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Ficha Del Taller",
        "description": "El taller y como le esta yendo, en una sola respuesta.\n\nSe abre antes de revisar orden por orden. Los dados de baja tambien se abren: cuando\nun taller se va, lo que uno quiere es entender por que.\n\nEs la unica ruta de este archivo que escribe, y lo que escribe es su propio rastro:\npara mirar cualquier cosa de un taller hay que pasar por aca primero, asi que con\nesta linea alcanza para saber quien entro.",
        "operationId": "ficha_del_taller_admin_workshops__taller_id__get",
        "parameters": [
          {
            "name": "taller_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Taller Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/workshops/{taller_id}/orders": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Ordenes Del Taller",
        "description": "Las ordenes del taller, con los mismos filtros que ve el taller en su panel.\n\nMismos filtros a proposito: cuando el dueno dice \"no me aparece la orden de la\ncamioneta\", hay que poder mirar exactamente lo que el esta mirando.",
        "operationId": "ordenes_del_taller_admin_workshops__taller_id__orders_get",
        "parameters": [
          {
            "name": "taller_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Taller Id"
            }
          },
          {
            "name": "open",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Open"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/workshops/{taller_id}/orders/{orden_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Detalle De La Orden",
        "description": "La orden con su bitacora y todos sus avisos.\n\nEs la pantalla del post-mortem: quien la movio, cuando, y si el WhatsApp de cada\ncambio llego a salir o quedo esperando que alguien apretara enviar. El panel del\ntaller muestra solo el ultimo aviso -es lo que necesita para abrir wa.me-, y con eso\nno se puede contestar \"a mi cliente nunca le avisaron\".",
        "operationId": "detalle_de_la_orden_admin_workshops__taller_id__orders__orden_id__get",
        "parameters": [
          {
            "name": "taller_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Taller Id"
            }
          },
          {
            "name": "orden_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Orden Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AltaTallerEntrada": {
        "properties": {
          "workshopName": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Workshopname"
          },
          "workshopPhone": {
            "type": "string",
            "maxLength": 20,
            "minLength": 8,
            "title": "Workshopphone"
          },
          "ownerName": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Ownername"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "workshopName",
          "workshopPhone",
          "ownerName",
          "email",
          "password"
        ],
        "title": "AltaTallerEntrada",
        "description": "Crea el taller y su usuario dueno de una sola vez."
      },
...
      "ClienteEntrada": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Name"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "rut": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rut"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "name",
          "phone"
        ],
        "title": "ClienteEntrada"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrdenEntrada": {
        "properties": {
          "clientId": {
            "type": "string",
            "title": "Clientid"
          },
          "vehicleId": {
            "type": "string",
            "title": "Vehicleid"
          },
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "estimatedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimatedat"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "clientId",
          "vehicleId",
          "title"
        ],
        "title": "OrdenEntrada"
      },
      "TallerEdicion": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20,
                "minLength": 8
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          }
        },
        "type": "object",
        "title": "TallerEdicion",
        "description": "Solo lo que se corrige a mano. El nombre sale en el WhatsApp que lee el cliente.\n\n`active` en falso suspende el taller: nadie de ahi entra, y sus datos quedan enteros."
      },
...
      "VehiculoEntrada": {
        "properties": {
          "clientId": {
            "type": "string",
            "title": "Clientid"
          },
          "plate": {
            "type": "string",
            "title": "Plate"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "clientId",
          "plate"
        ],
        "title": "VehiculoEntrada"
      }
    }
  }
}